+++ migration/index.spec.ts
@@ -18,7 +18,83 @@ describe("Migration", () => {
   });
 
   tester.run("v26", migration.rules.v26, {
-    valid: [`import {} from "zod";`],
-    invalid: [],
+    valid: [`const routing = { "get /": someEndpoint };`],
+    invalid: [
+      {
+        name: "basic DependsOnMethod",
+        code: `const routing = new DependsOnMethod({ get: someEndpoint });`,
+        output: `const routing = {\n"get /": someEndpoint,\n};`,
+        errors: [
+          {
+            messageId: "change",
+            data: {
+              subject: "value",
+              from: "new DependsOnMethod(...)",
+              to: "its argument object and append its keys with ' /'",
+            },
+          },
+        ],
+      },
+      {
+        name: "DependsOnMethod with literals",
+        code: `const routing = new DependsOnMethod({ "get": someEndpoint });`,
+        output: `const routing = {\n"get /": someEndpoint,\n};`,
+        errors: [
+          {
+            messageId: "change",
+            data: {
+              subject: "value",
+              from: "new DependsOnMethod(...)",
+              to: "its argument object and append its keys with ' /'",
+            },
+          },
+        ],
+      },
+      {
+        name: "deprecated DependsOnMethod",
+        code: `const routing = new DependsOnMethod({ get: someEndpoint }).deprecated();`,
+        output: `const routing = {\n"get /": someEndpoint.deprecated(),\n};`,
+        errors: [
+          {
+            messageId: "change",
+            data: {
+              subject: "value",
+              from: "new DependsOnMethod(...)",
+              to: "its argument object and append its keys with ' /'",
+            },
+          },
+        ],
+      },
+      {
+        name: "DependsOnMethod with nesting",
+        code: `const routing = new DependsOnMethod({ get: someEndpoint }).nest({ some: otherEndpoint });`,
+        output: `const routing = {\n"get /": someEndpoint,\n"some": otherEndpoint,\n};`,
+        errors: [
+          {
+            messageId: "change",
+            data: {
+              subject: "value",
+              from: "new DependsOnMethod(...)",
+              to: "its argument object and append its keys with ' /'",
+            },
+          },
+        ],
+      },
+      {
+        name: "DependsOnMethod both deprecated and with nesting",
+        code: `const routing = new DependsOnMethod({ get: someEndpoint }).deprecated().nest({ some: otherEndpoint });`,
+        output: `const routing = {\n"get /": someEndpoint.deprecated(),\n"some": otherEndpoint,\n};`,
+        errors: [
+          {
+            messageId: "change",
+            data: {
+              subject: "value",
+              from: "new DependsOnMethod(...)",
+              to: "its argument object and append its keys with ' /'",
+            },
+          },
+        ],
+      },
+    ],
   });
 });