+++ migration/index.ts
@@ -1,22 +1,32 @@
 import {
   ESLintUtils,
-  // AST_NODE_TYPES as NT,
+  AST_NODE_TYPES as NT,
   type TSESLint,
-  // type TSESTree,
+  type TSESTree,
 } from "@typescript-eslint/utils"; // eslint-disable-line allowed/dependencies -- assumed transitive dependency
 
-/*
 type NamedProp = TSESTree.PropertyNonComputedName & {
-  key: TSESTree.Identifier;
+  key: TSESTree.Identifier | TSESTree.StringLiteral;
 };
- */
 
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type -- temporary
-interface Queries {}
+interface Queries {
+  dependsOnMethod: TSESTree.NewExpression;
+}
 
 type Listener = keyof Queries;
 
-const queries: Record<Listener, string> = {};
+const queries: Record<Listener, string> = {
+  dependsOnMethod: `${NT.NewExpression}[callee.name='DependsOnMethod']`,
+};
+
+const isNamedProp = (prop: TSESTree.ObjectLiteralElement): prop is NamedProp =>
+  prop.type === NT.Property &&
+  !prop.computed &&
+  (prop.key.type === NT.Identifier ||
+    (prop.key.type === NT.Literal && typeof prop.key.value === "string"));
+
+const getPropName = (prop: NamedProp): string =>
+  prop.key.type === NT.Identifier ? prop.key.name : prop.key.value;
 
 const listen = <
   S extends { [K in Listener]: TSESLint.RuleFunction<Queries[K]> },
@@ -44,7 +54,58 @@ const v26 = ESLintUtils.RuleCreator.withoutDocs({
     },
   },
   defaultOptions: [],
-  create: () => listen({}),
+  create: (ctx) =>
+    listen({
+      dependsOnMethod: (node) => {
+        if (node.arguments.length !== 1) return;
+        const argument = node.arguments[0];
+        if (argument.type !== NT.ObjectExpression) return;
+        let isDeprecated = false;
+        let nested: TSESTree.ObjectExpression | undefined = undefined;
+        let cursor: TSESTree.Node = node;
+        while (
+          cursor &&
+          cursor.parent &&
+          cursor.parent.type === NT.MemberExpression &&
+          cursor.parent.property.type === NT.Identifier &&
+          cursor.parent.parent &&
+          cursor.parent.parent.type === NT.CallExpression
+        ) {
+          const name = cursor.parent.property.name;
+          const call = cursor.parent.parent as TSESTree.CallExpression;
+          if (name === "deprecated") isDeprecated = true;
+          if (
+            name === "nest" &&
+            call.arguments[0] &&
+            call.arguments[0].type === NT.ObjectExpression
+          )
+            nested = call.arguments[0];
+          cursor = call;
+        }
+        ctx.report({
+          node: cursor,
+          messageId: "change",
+          data: {
+            subject: "value",
+            from: "new DependsOnMethod(...)",
+            to: "its argument object and append its keys with ' /'",
+          },
+          fix: (fixer) => {
+            const makeMapper =
+              (feat?: "deprecated" | "nest") =>
+              (prop: TSESTree.ObjectLiteralElement) =>
+                isNamedProp(prop)
+                  ? `"${getPropName(prop)}${feat === "nest" ? "" : " /"}": ${ctx.sourceCode.getText(prop.value)}${feat === "deprecated" ? ".deprecated()" : ""},`
+                  : `${ctx.sourceCode.getText(prop)}, /** @todo migrate manually */`;
+            const nextProps = argument.properties
+              .map(makeMapper(isDeprecated ? "deprecated" : undefined))
+              .concat(nested?.properties.map(makeMapper("nest")) ?? [])
+              .join("\n");
+            return fixer.replaceText(cursor, `{\n${nextProps}\n}`);
+          },
+        });
+      },
+    }),
 });
 
 export default {